# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://test-project.test'
paths:
...
  '/api/contacts/{id}':
    get:
      summary: 'Contact Shows single .'
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  code: 200
                  locale: en
                  message: OK
                  data:
                    contact:
                      id: 1
                      account_id: '1'
                      first_name: Talon
                      last_name: Bahringer
                      job_title: 'Plating Machine Operator'
                      address: "94628 Craig Shoals\nHarveyfort, RI 01412-6293"
                      city: Abbyhaven
                      postal_code: 82793-5888
                      country_code: HK
                      email: isaiah.bogisich@gmail.com
                      phone: +1-351-936-2170
                      created_at: '2021-11-22T20:42:08.000000Z'
                      updated_at: '2021-11-22T20:42:08.000000Z'
                      account:
                        id: 1
                        name: 'Parisian PLC'
                        user_id: '1'
                        description: 'Minima quia quod natus. Enim aspernatur in iure doloribus distinctio aut fuga. Iste est distinctio recusandae enim molestias saepe. Sed omnis eligendi sunt.'
                        address: "618 Gulgowski Squares\nSouth Christophe, CA 47461"
                        city: 'South Dakota'
                        postal_code: 62172-5816
                        country_code: RS
                        is_active: '1'
                        created_at: '2021-11-22T20:42:08.000000Z'
                        updated_at: '2021-11-22T20:42:08.000000Z'
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: integer
                    example: 200
                  locale:
                    type: string
                    example: en
                  message:
                    type: string
                    example: OK
                  data:
                    type: object
                    example:
                      contact:
                        id: 1
                        account_id: '1'
                        first_name: Talon
                        last_name: Bahringer
                        job_title: 'Plating Machine Operator'
                        address: "94628 Craig Shoals\nHarveyfort, RI 01412-6293"
                        city: Abbyhaven
                        postal_code: 82793-5888
                        country_code: HK
                        email: isaiah.bogisich@gmail.com
                        phone: +1-351-936-2170
                        created_at: '2021-11-22T20:42:08.000000Z'
                        updated_at: '2021-11-22T20:42:08.000000Z'
                        account:
                          id: 1
                          name: 'Parisian PLC'
                          user_id: '1'
                          description: 'Minima quia quod natus. Enim aspernatur in iure doloribus distinctio aut fuga. Iste est distinctio recusandae enim molestias saepe. Sed omnis eligendi sunt.'
                          address: "618 Gulgowski Squares\nSouth Christophe, CA 47461"
                          city: 'South Dakota'
                          postal_code: 62172-5816
                          country_code: RS
                          is_active: '1'
                          created_at: '2021-11-22T20:42:08.000000Z'
                          updated_at: '2021-11-22T20:42:08.000000Z'
      tags:
        - Endpoints
    put:
      summary: 'Contact Updates an existing.'
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  code: 200
                  locale: en
                  message: OK
                  data:
                    contact:
                      id: 1
                      account_id: 1
                      first_name: Juvenal
                      last_name: Yost
                      job_title: 'Transformer Repairer'
                      address: "2615 Kaley Lodge\nEast Alfonsotown, CT 70458-6001"
                      city: Haagport
                      postal_code: '14234'
                      country_code: FO
                      email: newton43@keeling.com
                      phone: +1-541-245-6893
                      created_at: '2021-11-22T20:42:42.000000Z'
                      updated_at: '2021-11-22T20:42:42.000000Z'
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: integer
                    example: 200
                  locale:
                    type: string
                    example: en
                  message:
                    type: string
                    example: OK
                  data:
                    type: object
                    example:
                      contact:
                        id: 1
                        account_id: 1
                        first_name: Juvenal
                        last_name: Yost
                        job_title: 'Transformer Repairer'
                        address: "2615 Kaley Lodge\nEast Alfonsotown, CT 70458-6001"
                        city: Haagport
                        postal_code: '14234'
                        country_code: FO
                        email: newton43@keeling.com
                        phone: +1-541-245-6893
                        created_at: '2021-11-22T20:42:42.000000Z'
                        updated_at: '2021-11-22T20:42:42.000000Z'
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: gd
                last_name:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: e
                job_title:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: ''
                account_id:
                  type: integer
                  description: ''
                  example: 19
                postal_code:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: ki
                city:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: cp
                address:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: i
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: kshlerin.gustave@example.net
                country_code:
                  type: string
                  description: ''
                  example: null
              required:
                - first_name
                - last_name
                - account_id
    delete:
      summary: 'Contact Deletes an existing.'
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  code: 200
                  locale: en
                  message: OK
                  data:
                    value: 'Contact successfully deleted'
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: integer
                    example: 200
                  locale:
                    type: string
                    example: en
                  message:
                    type: string
                    example: OK
                  data:
                    type: object
                    example:
                      value: 'Contact successfully deleted'
      tags:
        - Endpoints
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the contact.'
        example: 17
        required: true
        schema:
          type: integer
...
  /api/contacts:
    get:
      summary: 'Contacts Returns all.'
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  code: 200
                  locale: en
                  message: OK
                  data:
                    contacts:
                      current_page: 1
                      data: []
                      first_page_url: 'http://example.test/api/contacts?page=1'
                      from: null
                      last_page: 1
                      last_page_url: 'http://example.test/api/contacts?page=1'
                      links:
                        -
                          url: null
                          label: '&laquo; Previous'
                          active: false
                        -
                          url: 'http://example.test/api/contacts?page=1'
                          label: '1'
                          active: true
                        -
                          url: null
                          label: 'Next &raquo;'
                          active: false
                      next_page_url: null
                      path: 'http://example.test/api/contacts'
                      per_page: 10
                      prev_page_url: null
                      to: null
                      total: 0
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: integer
                    example: 200
                  locale:
                    type: string
                    example: en
                  message:
                    type: string
                    example: OK
                  data:
                    type: object
                    example:
                      contacts:
                        current_page: 1
                        data: []
                        first_page_url: 'http://example.test/api/contacts?page=1'
                        from: null
                        last_page: 1
                        last_page_url: 'http://example.test/api/contacts?page=1'
                        links:
                          -
                            url: null
                            label: '&laquo; Previous'
                            active: false
                          -
                            url: 'http://example.test/api/contacts?page=1'
                            label: '1'
                            active: true
                          -
                            url: null
                            label: 'Next &raquo;'
                            active: false
                        next_page_url: null
                        path: 'http://example.test/api/contacts'
                        per_page: 10
                        prev_page_url: null
                        to: null
                        total: 0
      tags:
        - Endpoints
    post:
      summary: 'Contacts Stores newly created.'
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  code: 201
                  locale: en
                  message: OK
                  data:
                    contact:
                      id: 1
                      first_name: Adelia
                      last_name: Legros
                      account_id: 1
                      postal_code: 43201-4839
                      city: Kiehnfurt
                      address: "4288 Amalia Pine Suite 107\nAndreannemouth, KS 14490-1489"
                      email: pfeffer.angus@cassin.com
                      country_code: VI
                      updated_at: '2021-11-22T20:40:56.000000Z'
                      created_at: '2021-11-22T20:40:56.000000Z'
                properties:
                  success:
                    type: boolean
                    example: true
                  code:
                    type: integer
                    example: 201
                  locale:
                    type: string
                    example: en
                  message:
                    type: string
                    example: OK
                  data:
                    type: object
                    example:
                      contact:
                        id: 1
                        first_name: Adelia
                        last_name: Legros
                        account_id: 1
                        postal_code: 43201-4839
                        city: Kiehnfurt
                        address: "4288 Amalia Pine Suite 107\nAndreannemouth, KS 14490-1489"
                        email: pfeffer.angus@cassin.com
                        country_code: VI
                        updated_at: '2021-11-22T20:40:56.000000Z'
                        created_at: '2021-11-22T20:40:56.000000Z'
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: 'n'
                last_name:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: ny
                job_title:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: af
                account_id:
                  type: integer
                  description: ''
                  example: 10
                postal_code:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: wm
                city:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: t
                address:
                  type: string
                  description: 'Must be at least 2 characters.'
                  example: ''
                email:
                  type: string
                  description: 'Must be a valid email address.'
                  example: xgreenfelder@example.org
                country_code:
                  type: string
                  description: ''
                  example: null
              required:
                - first_name
                - last_name
                - account_id
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []